+++ elasticsearch/scripts/elasticsearch-custom-entrypoint.sh
@@ -1,22 +1,26 @@
 #!/bin/bash -e
 
 # This script is using the given ELASTICSEARCH_HOSTS to populate required parameters
-# to setup a Multi-Node Elasticsearch cluster.
-# If parameters (network.publish_host, http.port, transport.port) are already 
+# to setup a Multi-Node Elasticsearch cluster. It is done with the help of the 
+# environment variable node.name which is set by Elasticsearch itself within the container.
+# If parameters (network.publish_host, http.port, transport.port) are  
 # given externally they are used with precedence.
 
-# Get the number of the Elasticsearch node based on the node.name (e.g. elasticsearch2)
-echo "Setup Elasticsearch cluster based on ELASTICSEARCH_HOSTS: ${ELASTICSEARCH_HOSTS}"
-# Translate the dotted environment variable into Non-Dotted to avoid issues with Bash commands
+# It is required to translate the dotted environment variable into Non-Dotted 
+# to make is accessible with bash commands
 nodeName=`echo | awk '{ print ENVIRON["node.name"] }'`
 
+# Get the number and basename of the Elasticsearch node based on the node.name (e.g. elasticsearch2)
 nodeBasename=`echo $nodeName | sed -r 's/([a-zA-Z]*)([0-9]{1,})/\1/'`
 nodeNumber=`echo $nodeName | sed -r 's/([a-zA-Z]*)([0-9]{1,})/\2/'`
+
+echo "Setup Elasticsearch cluster node: ${nodeNumber} based on ELASTICSEARCH_HOSTS: ${ELASTICSEARCH_HOSTS}"
 
 count=1
 params=""
 seedHosts=""
 initialMasterNode=""
+nodeConfigured=false
 for host in ${ELASTICSEARCH_HOSTS//,/ }
 do
     # Use only the first node as initial master node, when initializing the cluster
@@ -25,8 +29,8 @@ do
         echo "Init Elasticsearch cluster using nodeBasename: ${nodeBasename} and count: ${count}"
         initialMasterNode="-E cluster.initial_master_nodes=${nodeBasename}${count}"
     fi
-    # Use all declared hosts as seed hosts if 
-    # Seeds hosts are not given externally and the standard transport ports are used
+    # Use all declared hosts as seed hosts, but only if
+    # seed hosts are not given externally and the standard transport ports are used
     discoverySeedHosts=`echo | awk '{ print ENVIRON["discovery.seed_hosts"] }'`
     discoveryTransportPort=`echo | awk '{ print ENVIRON["transport.port"] }'`
     if [ -z "${discoverySeedHosts}" -a -z "${discoveryTransportPort}" ]
@@ -42,7 +46,7 @@ do
     fi
 
     if [ "${count}" == "${nodeNumber}" ]; then
-        echo "Setting up Elasticsearch node: ${nodeNumber}"
+        nodeConfigured=true
         publishHost=`echo | awk '{ print ENVIRON["network.publish_host"] }'`
         if [ -z "${publishHost}" ]; then
             publishHost=`echo $host | sed -r 's/https?:\/\/(.*)\:[0-9]{4}/\1/'`
@@ -72,8 +76,14 @@ do
             echo "transport.port=${transportPort} taken from envionment variable"
         fi
     fi
-    count=`expr $count + 1`    
+    count=`expr $count + 1`
 done
+
+if [ ${nodeConfigured} == false ]; then
+    echo "Failed to start/configure Elasticsearch node: ${nodeNumber}. Please check that ELASTICSEARCH_HOSTS contains the required number of nodes.";
+    exit 99
+fi
+exit 1
 
 # Check if Self-Monitoring should be enabled
 if [ -z "${SELF_MONITORING_ENABLED}" ];then